+++ src/main/java/sid/roborally/game_mechanics/grid/Hole.java
@@ -1,4 +1,4 @@
-package sid.roborally.game_mechanics;
+package sid.roborally.game_mechanics.grid;
 
 public class Hole extends GridObject {
     public Hole(int x, int y) {
